# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order API",
        "description": "API для управління замовленнями",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизація користувача",
                "operationId": "535ffeda4b78916efd658b2844ea897a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успішна авторизація",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Некоректні облікові дані"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Отримати дані про поточного користувача",
                "operationId": "1537d2a532022d0974e4cd2d83c6a38f",
                "responses": {
                    "200": {
                        "description": "Інформація про користувача",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID користувача",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Ім'я користувача",
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "description": "Електронна пошта користувача",
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Вихід з системи",
                "operationId": "46f8a14c8083abb69eed72768aa08629",
                "responses": {
                    "200": {
                        "description": "Успішний вихід",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Оновлення токену доступу",
                "operationId": "f3538b6e9c46bcf601a9d194288b03d4",
                "responses": {
                    "200": {
                        "description": "Новий токен доступу",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Реєстрація нового користувача",
                "operationId": "da8643013c2941f5ed7f8a9c9b4a7f7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успішна реєстрація",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Користувач успішно зареєстрований"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user@example.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Отримати список замовлень",
                "operationId": "22d21713fd1e12635d9297e93c85a1ba",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер сторінки для пагінації",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Поле для сортування",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Напрямок сортування (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список замовлень",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Створити нове замовлення",
                "operationId": "e032a80fdff957729ab78daa843e6bc2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Замовлення успішно створено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Отримати замовлення за ID",
                "operationId": "5372efb35716a5574a7cf40aa8392116",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID замовлення",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деталі замовлення",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Замовлення не знайдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Оновити замовлення",
                "operationId": "35e607c39383af6af927a0a4dd0104cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID замовлення",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Замовлення успішно оновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Видалити замовлення",
                "operationId": "ab5ef85086407a6f38022461a4f02c51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID замовлення",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Замовлення успішно видалено"
                    },
                    "404": {
                        "description": "Замовлення не знайдено"
                    }
                }
            }
        },
        "/api/orders/{id}/change-status": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Змінити статус замовлення",
                "operationId": "a8ebb31ac58fd431bb79d05e1e704416",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID замовлення",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeOrderStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Статус замовлення успішно змінено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Замовлення не знайдено"
                    }
                }
            }
        },
        "/api/orders/export-csv": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Експортувати замовлення у CSV",
                "operationId": "584d2c24fd8ec75bf87da6758e51e881",
                "responses": {
                    "200": {
                        "description": "CSV файл успішно створено"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Електронна пошта користувача",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "description": "Пароль користувача",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "description": "Ім'я користувача",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "Електронна пошта користувача",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "description": "Пароль користувача",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "description": "Підтвердження пароля",
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "TokenResponse": {
                "properties": {
                    "access_token": {
                        "description": "JWT токен доступу",
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR..."
                    },
                    "token_type": {
                        "description": "Тип токену",
                        "type": "string",
                        "example": "bearer"
                    },
                    "expires_in": {
                        "description": "Термін дії токену (в секундах)",
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "Order": {
                "required": [
                    "product_name",
                    "amount",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_name": {
                        "type": "string",
                        "example": "Example Product"
                    },
                    "amount": {
                        "type": "number",
                        "example": 150.75
                    },
                    "status": {
                        "type": "string",
                        "example": "new"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "StoreOrderRequest": {
                "required": [
                    "product_name",
                    "amount"
                ],
                "properties": {
                    "product_name": {
                        "description": "Назва продукту",
                        "type": "string",
                        "example": "Example Product"
                    },
                    "amount": {
                        "description": "Сума замовлення",
                        "type": "number",
                        "example": 100.5
                    }
                },
                "type": "object"
            },
            "UpdateOrderRequest": {
                "required": [],
                "properties": {
                    "product_name": {
                        "description": "Оновлена назва продукту",
                        "type": "string",
                        "example": "Updated Product"
                    },
                    "amount": {
                        "description": "Оновлена сума замовлення",
                        "type": "number",
                        "example": 200.75
                    },
                    "status": {
                        "description": "Статус замовлення",
                        "type": "string",
                        "example": "completed"
                    }
                },
                "type": "object"
            },
            "ChangeOrderStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "description": "Новий статус замовлення",
                        "type": "string",
                        "example": "completed"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API для аутентифікації"
        },
        {
            "name": "Orders",
            "description": "API для роботи з замовленнями"
        }
    ]
}